// src/types/TimesType.ts
export type TimesTypes = {
  region: string;
  date: "04/03/2024, 01:00:00";
  hijri_date: {
    month: "sha'bon";
    day: 23;
  };
  weekday: string;
  times: {
    tong_saharlik: "05:37";
    quyosh: "06:53";
    peshin: "12:35";
    asr: "16:33";
    shom_iftor: "18:22";
    hufton: "19:33";
  };
};